+++ wetterdienst/core/scalar/tools.py
@@ -16,7 +16,6 @@ def extract_station_values(
     # 1. only add further stations if not a minimum of 4 stations is reached OR
     # 2. a gain of 10% of timestamps with at least 4 existing values over all stations is seen OR
     # 3. an additional counter is below 3 (used if a station has really no or few values)
-
     cond1 = param_data.values.shape[1] < 4
     cond2 = not cond1 and gain_of_value_pairs(param_data.values, result_series_param) > 0.10
     if (